+++ RELEASE_NOTES.md
@@ -6,8 +6,11 @@ This is the offical release of vpinfe v1.1.38.
 ### What's New
 - updated the theme doc to account for the new bg and dmd high DPI overrides
 - added "Score" section to .info
+- added non-rom score parsing
+- added initals as part of vpinplay sync 
 
 ### Fixes
 None
+
 ### Notes
 None